Add build testing with GitHub actions
Currently the project doesn't build with IDF versions v4.4 and v5.0 and
we're not aware about that:

 mipi_display.c: In function 'mipi_display_init':
 mipi_display.c:204:9: error: implicit declaration of function 'esp_rom_gpio_pad_select_gpio'; did you mean 'gpio_pad_select_gpio'? [-Werror=implicit-function-declaration]
         esp_rom_gpio_pad_select_gpio(CONFIG_MIPI_DISPLAY_PIN_CS);
         ^~~~~~~~~~~~~~~~~~~~~~~~~~~~
         gpio_pad_select_gpio

So lets improve the UX and add simple build QA using CI build workflow
based on official Espressif's IDF SDK containers.

For the start simply build every sdkconfig configuration with v4.4 and
v5.0 of IDF SDKs.

Signed-off-by: Petr Štetiar <ynezz@true.cz>